wayland: remove weston-conf.bbappend
Meta-IVI technically uses the poky repo as its upstream source for
OE-core, and it was most recently based on poky/thud. Meta-ivi has a
bbappend on weston-conf.bb, which exists in poky - but is out of date -
and *used* to exist in OE-core, but has since been deprecated in favor
of weston-init. NILRT doesn't provide poky in its layer stack, and so
there is nothing for the weston-conf.bbappend to apply to.

Upstream meta-ivi needs to fix their wayland configuration. But since
NILRT doesn't support wayland, we can just drop the bbappend to keep OE
from warning us about it being baseless.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>